+++ src/random/randomInt.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, test } from "@jest/globals";
-import { randomInt } from "./randomInt";
+import { randomBigInt, randomInt } from "./randomInt";
 
 describe("randomInt", () => {
   test("no args", async () => {
@@ -11,5 +11,9 @@ describe("randomInt", () => {
     expect(randomInt({ min: 12, max: 20 })).toBeGreaterThanOrEqual(12);
     expect(randomInt({ min: 12, max: 12 })).toBeLessThanOrEqual(12);
     expect(randomInt({ min: 11, max: 12 })).toBeLessThanOrEqual(13);
+  });
+
+  test("randomBigInt", async () => {
+    expect(randomBigInt().toString()).not.toContain("n");
   });
 });