+++ src/random/randomValue.ts
@@ -1,8 +1,9 @@
 import { randomArrayItem } from "./randomArrayItem";
 import { randomBool } from "./randomBool";
 import { randomDate } from "./randomDate";
-import { randomInt } from "./randomInt";
+import { randomBigInt, randomInt } from "./randomInt";
 import { randomString } from "./randomString";
+import { randomSymbol } from "./randomSymbol";
 
 export const randomValue = () => {
   return randomArrayItem([
@@ -10,7 +11,11 @@ export const randomValue = () => {
     randomString(),
     randomInt(),
     randomDate(),
+    randomBigInt(),
+    randomSymbol(),
     undefined,
     null,
+    NaN,
+    Infinity,
   ]);
 };